Refactor ConsensusStateHeightQuerier
