+++ crates/relayer-cosmos/src/methods/mod.rs
@@ -1,5 +1,4 @@
 pub mod channel;
-pub mod consensus_state;
 pub mod encode;
 pub mod event;
 pub mod packet;
